Shares containers accross all tests
Adds a health test, too
